+++ backend/src/services/cache.rs
@@ -264,7 +264,7 @@ impl CacheService {
         
         match ttl {
             Some(duration) => {
-                conn.set_ex::<_, _, ()>(key, serialized, duration.as_secs())
+                conn.set_ex::<_, _, ()>(key, serialized, duration.as_secs() as usize)
                     .map_err(|e| AppError::InternalServerError(format!("Redis set_ex error: {}", e)))?;
             }
             None => {
@@ -355,7 +355,7 @@ impl CacheService {
         
         if let Some(duration) = ttl {
             for key in serialized_values.keys() {
-                conn.expire::<_, ()>(key, duration.as_secs() as i64)
+                conn.expire::<_, ()>(key, duration.as_secs() as usize)
                     .map_err(|e| AppError::InternalServerError(format!("Redis expire error: {}", e)))?;
             }
         }
